Feat: Move the dataset permission drop-down box to a separate file for better permission control #3221
