+++ web/src/pages/dataset/setting/general-form.tsx
@@ -1,7 +1,5 @@
 import { AvatarUpload } from '@/components/avatar-upload';
 import { FormContainer } from '@/components/form-container';
-import { SelectWithSearch } from '@/components/originui/select-with-search';
-import { RAGFlowFormItem } from '@/components/ragflow-form';
 import { Button } from '@/components/ui/button';
 import {
   FormControl,
@@ -11,22 +9,14 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
-import { PermissionRole } from '@/constants/permission';
-import { useMemo } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
+import { PermissionFormField } from './permission-form-field';
 import { GeneralSavingButton } from './saving-button';
 
 export function GeneralForm() {
   const form = useFormContext();
   const { t } = useTranslation();
-
-  const teamOptions = useMemo(() => {
-    return Object.values(PermissionRole).map((x) => ({
-      label: t('knowledgeConfiguration.' + x),
-      value: x,
-    }));
-  }, [t]);
 
   return (
     <>
@@ -98,17 +88,7 @@ export function GeneralForm() {
             );
           }}
         />
-        <RAGFlowFormItem
-          name="permission"
-          label={t('knowledgeConfiguration.permissions')}
-          tooltip={t('knowledgeConfiguration.permissionsTip')}
-          horizontal
-        >
-          <SelectWithSearch
-            options={teamOptions}
-            triggerClassName="w-3/4"
-          ></SelectWithSearch>
-        </RAGFlowFormItem>
+        <PermissionFormField></PermissionFormField>
       </FormContainer>
       <div className="text-right pt-4 flex justify-end gap-3">
         <Button
